Create ColorPalette-WSJ_Infectious_Disease
The Wall Street Journal recently posted a nice data visualization on [the impact of vaccines in the United States](http://graphics.wsj.com/infectious-diseases-and-vaccines/). I really liked the color palette they used, and wanted to recreate it for Tableau. Add the code to your preferences.tps file in your My Tableau Repository folder.